services/products/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
PyJWT
python-dotenv
